fix: todos and some perf
* fix: MAX_CHANGE_RATE
* test: reentrancy for mint and burn
* lint: disable solhint for empty block
* perf: using `uint256` to load reserves instead of `uint104`
* perf: remove comment since it is already done
* perf: remove comment as it is tested and found to be the same
* perf: replaced own math lib with more optimized ones
* perf: remove comments
* perf: make _calcFee unchecked
* fix: import path typo
* fix: rename variable in tests as well
* fix: variable naming convention
* fix: more variable naming convention
* fix: remove comment
* coverage: update key
* fix: variable name
* fix: removing unnecessary use and casting of uint104
* fix: removed unused casting
* feat: use SafeTransferLib for safeTransferFrom
* fix: substitute variable
* fix: substitute variable
* fix: use safeTransfer instead of Transfer for adjustManagement
* fix: use safeApprove instead of approve
* test: use sqrt implementation for tests
* test: demonstrate problem of managed amount greater than uint104
* feat: SafeCast for adjustManagement
* test: adjustManagement uint104
* fix: removed unused `require` statement
* test: update gas snapshot

+++ src/interfaces/IAssetManager.sol
@@ -6,7 +6,7 @@ import { ERC20 } from "solmate/tokens/ERC20.sol";
 import { ReservoirPair } from "src/ReservoirPair.sol";
 
 interface IAssetManager {
-    function getBalance(ReservoirPair owner, ERC20 token) external returns (uint104 tokenBalance);
+    function getBalance(ReservoirPair owner, ERC20 token) external returns (uint256 tokenBalance);
     function afterLiquidityEvent() external;
     function returnAsset(bool aToken0, uint256 aAmount) external;
 }
